+++ lib/object_storage/config.rb
@@ -84,11 +84,14 @@ module ObjectStorage
 
     def fog_attributes
       @fog_attributes ||= begin
-        return {} unless enabled? && aws?
-        return {} unless server_side_encryption.present?
+        return {} unless aws_server_side_encryption_enabled?
 
         aws_server_side_encryption_headers.compact
       end
+    end
+
+    def aws_server_side_encryption_enabled?
+      aws? && server_side_encryption.present?
     end
 
     private