+++ spec/finders/ci/pipelines_finder_spec.rb
@@ -258,20 +258,8 @@ RSpec.describe Ci::PipelinesFinder do
       let!(:push_pipeline) { create(:ci_pipeline, project: project, source: 'push') }
       let!(:api_pipeline) { create(:ci_pipeline, project: project, source: 'api') }
 
-      context 'when `pipeline_source_filter` feature flag is disabled' do
-        before do
-          stub_feature_flags(pipeline_source_filter: false)
-        end
-
-        it 'returns all the pipelines' do
-          is_expected.to contain_exactly(web_pipeline, push_pipeline, api_pipeline)
-        end
-      end
-
-      context 'when `pipeline_source_filter` feature flag is enabled' do
-        it 'returns only the matched pipeline' do
-          is_expected.to eq([web_pipeline])
-        end
+      it 'returns only the matched pipeline' do
+        is_expected.to eq([web_pipeline])
       end
     end